Implements tRCP for event types
